requirements.txt
langchain-ollama
ollama
langchain-openai
python-dotenv